backend/requirements.txt
# ============================================
# Logistics API Platform - Python Dependencies
# ============================================

# Web Framework
Flask==3.1.1
flask-cors==5.0.1

# Database
pyodbc==5.2.0
SQLAlchemy==2.0.40

# Security (Werkzeug and PyJWT)
python-dotenv==1.1.0
PyJWT==2.10.1

# Utilities
marshmallow==3.26.1
openpyxl==3.1.5
requests==2.32.3
